Integrate validation constraints and enhance functionality
Added 'NotBlank' validation constraints to the 'name' and 'type' fields in the FormField and Form classes. Created a construct function in the Builder class and method to build forms from definitions. Set up a service for form factory in services config and added 'DuplicatedFieldName' validator to avoid duplicate field names in forms.